Bugfixes for Edge, IE, Safari
Added a "dummy class" .st12 to the <tspan> children of <text>, in order to give it compatibility with Edge/IE/Safari.
Also added a new JavaScript function that detects the browser used - when it's NOT one of the previously listed, it eliminates the dummy class.

For More Info check the /* comment */ at line 914

Refactored some of the code, fixed some code bugs, renamed some classes with better names, rewrote some comments.